test: keep deck import accessibility test local
